[CIDMDEV-5130] - Migration Support - Read-Only Capabilities
- Protected the following services by flag 'feature.migration.read.only.mode.enabled'
(returns 503 when read-only enabled) :
1. PUT v2.0/RAX-AUTH/service-apis/{type}/{version}/capabilities
2. DELETE v2.0/RAX-AUTH/service-apis/{type}/{version}/capabilities

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -875,7 +875,11 @@ public class Cloud20VersionResource {
             @PathParam("type") String type,
             @PathParam("version") String version,
             Capabilities capabilities){
-        return cloud20Service.updateCapabilities(authToken, capabilities, type, version).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.updateCapabilities(authToken, capabilities, type, version).build();
+        }
     }
 
     @DELETE
@@ -884,7 +888,11 @@ public class Cloud20VersionResource {
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("type") String type,
             @PathParam("version") String version){
-        return cloud20Service.removeCapabilities(authToken, type, version).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.removeCapabilities(authToken, type, version).build();
+        }
     }
 
     @GET
